fix: end-to-end sync working against demo bank
Major fixes to get the full sync pipeline operational:

- Use functools.partial instead of lambdas in workflow.py so Restate
  detects async callables correctly (was serializing coroutine objects)
- Switch heritage adapter from human_fill/bezier clicks to page.fill/click
  (SPA form submissions stalled with character-by-character typing)
- Add SPA rendering waits: wait for table/content selectors instead of
  just networkidle (demo bank renders asynchronously from localStorage)
- Deterministic account navigation via tr:has(td) locator instead of
  LLM-based link finding (LLM incorrectly returned "done" on summary page)
- Increase max_tokens to 8192 for transaction extraction (was truncating
  at 1024, losing the closing bracket of large JSON arrays)
- Make _ask() max_tokens configurable per extraction goal
- Use Decimal for all money fields (Pydantic models + DB columns)
- Batch transaction inserts with ON CONFLICT DO NOTHING
- Fix playwright_stealth API (Stealth class, apply to context not page)
- Fix structlog config (remove add_logger_name, incompatible with PrintLogger)
- CLI: reuse existing BankConnection, async /run/send trigger, connection reuse
- Add tests for crypto, models, adapters, extractor

Verified: 3 accounts, 75 transactions, 3 balance snapshots synced from
https://demo-bank-2.vercel.app

+++ src/db/models.py
@@ -16,6 +16,7 @@ Designed for future horizontal scaling:
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import Any
 
 from sqlalchemy import (
@@ -140,8 +141,8 @@ class Balance(Base):
     account_id: Mapped[str] = mapped_column(
         UUID(as_uuid=False), ForeignKey("accounts.id"), nullable=False, index=True
     )
-    available: Mapped[float | None] = mapped_column(Numeric(20, 4))
-    current: Mapped[float] = mapped_column(Numeric(20, 4), nullable=False)
+    available: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
+    current: Mapped[Decimal] = mapped_column(Numeric(20, 4), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), nullable=False)
     captured_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
 
@@ -163,9 +164,9 @@ class Transaction(Base):
     external_id: Mapped[str] = mapped_column(Text, nullable=False)  # bank's own transaction ID
     posted_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), index=True)
     description: Mapped[str | None] = mapped_column(Text)
-    amount: Mapped[float] = mapped_column(Numeric(20, 4), nullable=False)  # negative = debit
+    amount: Mapped[Decimal] = mapped_column(Numeric(20, 4), nullable=False)  # negative = debit
     currency: Mapped[str] = mapped_column(String(3), nullable=False)
-    running_balance: Mapped[float | None] = mapped_column(Numeric(20, 4))
+    running_balance: Mapped[Decimal | None] = mapped_column(Numeric(20, 4))
     raw: Mapped[dict[str, Any] | None] = mapped_column(JSON)  # original scraped payload
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now(), nullable=False
